2차 인증 - 화이트리스트에 포함되는지

+++ src/main/java/com/example/sb_security_jwt/app/security/jwt/JwtProvider.java
@@ -19,7 +19,7 @@ public class JwtProvider {
         return jwtSecretKey;
     }
 
-    public String generateAccessToken(Map<String, Object> claims, int seconds) {
+    public String generateAccessToken(Map<String, Object> claims, long seconds) {
         long now = new Date().getTime();
         Date accessTokenExpiresIn = new Date(now + 1000L * seconds);
